[BUGFIX] Use mount point parameter in page document ID
The mount parameter that is used to access a page is now passed on
to the page indexer and will be added to the page document ID if
it is not empty. This allows a page to appear multiple times in the
index in its different mount point contexts.

Relates: #59438
Change-Id: I4fd0d4f21b1805fc532eef1a51851ddc8fa303ee